// src/app/models/Role.ts
export enum Role {

  ROLE_PROFISSIONAL_UBS = "ROLE_PROFISSIONAL_UBS",
	ROLE_AGENTE_ENDEMIAS = "ROLE_AGENTE_ENDEMIAS",
	ROLE_ADMINISTRADOR_ESTADUAL = "ROLE_ADMINISTRADOR_ESTADUAL",
	ROLE_ADMINISTRADOR_MUNICIPAL = "ROLE_ADMINISTRADOR_MUNICIPAL",
	ROLE_ADMINISTRADOR_GERAL = "ROLE_ADMINISTRADOR_GERAL",
	ROLE_ADMINISTRADOR_NACIONAL = "ROLE_ADMINISTRADOR_NACIONAL",
	ROLE_UBS = "ROLE_UBS",

}